requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
yarl==1.9.4
